requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
ecdsa==0.19.0
gunicorn==23.0.0
packaging==24.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.8.2
